Introduce BaseWithoutValidation
